hubspot_integration/requirements.txt
uvicorn==0.24.0
python-dotenv==1.0.0
hubspot-api-client==7.4.0
mcp>=1.6.0 